+++ utils/logger.py
@@ -1,0 +1,3 @@
+import logging
+
+logger: logging.Logger = logging.getLogger("logger_mco_main")
